+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New extra install `-E websockets`, additionally a convenience `-E all` option
 - `context_hook` in Base Manager that can be used to bind structlog contextvars for all workers (inbound, outbound, poll)
 - `connect_hook` and `disconnect_hook` for Websocket manager
+- `callback_hook` in Base Manager that's called just before a `QueuedMessage` is delegated to a callback, primarily for adding `topic` to logs
+- `JupyterKernelManager` picks up the same `last_seen_message` debug pattern that `WebsocketManager` has 
 
 ### Changed
 - Use `managed_service_fixtures` for Redis tests